feat: move Card component into its own component file
